html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/form-submit-popup-before-opener-load.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>

<form action="resources/message-opener.html"></form>
<img src="/common/slow.py">

<script>
window.onload = () => {
  window.onloadFired = true;
};

"use strict";
promise_test(async t => {
  const form = document.querySelector("form");

  const windowName = token();
  form.target = windowName;

  const code = `
    window.onload = () => { window.onloadFired = true; };
  `;
  const startURL = "resources/slow-code-injector.html?pipe=sub(none)&code=" + encodeURIComponent(code);
  const absoluteStartURL = (new URL(startURL, location.href)).href;
  const w = window.open(startURL, windowName);
  t.add_cleanup(() => w.close());

  // Wait to get past any initial about:blank
  while (true) {
    if (w.location.href === absoluteStartURL) {
      break;
    }
    await new Promise(r => setTimeout(r, 0));
  }

  assert_equals(window.onloadFired, undefined, "onload must not yet have fired in the opener window");
  assert_equals(w.onloadFired, undefined, "onload must not yet have fired in the opened window");
  assert_equals(w.history.length, 1, "history.length for the opened window must start at 1");

  form.submit();

  await new Promise(r => { window.onmessage = r; });
  assert_equals(w.history.length, 1, "history.length must not increase");
  assert_equals(w.location.href, form.action + "?");
}, "Replace before load, triggered by formElement.submit() in the opener window, before the opener has loaded");
</script>
